// src/models/loggerconfig.model.ts
export class LoggerConfig {
  LOGGERCONNECTIONSTRING?: string | null | undefined;
  SERVERIP: string | null | undefined;
  SERVERNAME: string | null | undefined;
  APPLICATIONNAME: string | null | undefined;
  LOGGERTYPE: number | null | undefined;
  DEBUG: boolean | null | undefined;
  AWSACCESSKEYID?: string | null | undefined;
  AWSSECRETKEY?: string | null | undefined;
  AWSREGION?: string | null | undefined;
}